# ai-specs/specs/api-spec.yml
openapi: "3.1.0"
info:
  title: Voice AI Runtime API
  version: "1.0.0"
  description: >
    API for the Voice AI Runtime — health checks, Prometheus metrics,
    and admin endpoints for call inspection.

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /health:
    get:
      summary: Health check
      description: >
        Returns system health status. Checks asyncpg pool connectivity,
        Redis reachability, and embedding model loading state.
      tags: [health]
      responses:
        "200":
          description: All checks passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: One or more checks failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /metrics:
    get:
      summary: Prometheus metrics
      description: >
        Returns all registered Prometheus metrics in text exposition format.
        Metrics include voice_turn_latency_ms, voice_route_a_confidence,
        voice_route_b_confidence, voice_tool_execution_ms, voice_barge_in_total,
        voice_fallback_llm_total, voice_active_calls, voice_filler_emitted_total.
      tags: [health]
      responses:
        "200":
          description: Prometheus metrics in text format
          content:
            text/plain:
              schema:
                type: string

  /api/v1/calls:
    get:
      summary: List recent calls
      description: Returns the 50 most recent call sessions ordered by started_at descending.
      tags: [admin]
      responses:
        "200":
          description: List of call sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallSummary"
        "503":
          description: Database not configured

    post:
      summary: Create voice call session
      description: >
        Creates a new voice call session. Returns a call_id used for
        subsequent SDP exchange. Enforces max_concurrent_calls limit.
      tags: [webrtc]
      responses:
        "201":
          description: Call session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCallResponse"
        "503":
          description: Max concurrent calls exceeded

  /api/v1/calls/{call_id}:
    get:
      summary: Get call detail
      description: >
        Returns a single call session with its turns and agent generations.
      tags: [admin]
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Call detail with turns and generations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallDetail"
        "404":
          description: Call not found
        "503":
          description: Database not configured

    delete:
      summary: End voice call
      description: >
        Ends an active voice call and removes it from the session registry.
      tags: [webrtc]
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Call ended and removed from registry
        "404":
          description: Call not found

  # ---- WebRTC Signaling ----

  /api/v1/calls/{call_id}/offer:
    post:
      summary: Two-step SDP exchange
      description: >
        Performs a two-step SDP exchange with OpenAI's Realtime WebRTC API.
        Step 1: POST /v1/realtime/sessions with session config (model, transcription,
        turn_detection with create_response=false) to obtain an ephemeral key.
        Step 2: POST /v1/realtime with the ephemeral key and SDP offer to get
        the SDP answer. The server API key is only used for Step 1 — never
        exposed to the browser. The browser connects directly to OpenAI for audio.
      tags: [webrtc]
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SDPRequest"
      responses:
        "200":
          description: SDP answer from OpenAI
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SDPResponse"
        "404":
          description: Call not found
        "502":
          description: OpenAI session creation or SDP exchange error

  /api/v1/calls/{call_id}/events:
    get:
      summary: Event forwarding WebSocket
      description: >
        WebSocket endpoint for bidirectional event forwarding between the
        browser and the Coordinator. Input: browser forwards OpenAI data channel
        events to the bridge for translation and dispatch. Output: Coordinator
        commands are sent to the browser for forwarding to the OpenAI data channel.
        On connection, a one-time session.update is sent to configure transcription
        (whisper-1) and disable auto-response (create_response=false).
      tags: [webrtc]
      parameters:
        - name: call_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "101":
          description: WebSocket upgrade accepted
        "4004":
          description: Call not found (WebSocket close code)

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded]
        checks:
          type: object
          properties:
            database:
              type: string
            redis:
              type: string
            models:
              type: string

    CallSummary:
      type: object
      properties:
        call_id:
          type: string
          format: uuid
        started_at:
          type: integer
          description: Epoch milliseconds
        ended_at:
          type: integer
          nullable: true
        status:
          type: string
        locale_hint:
          type: string
          nullable: true

    CallDetail:
      type: object
      properties:
        call:
          $ref: "#/components/schemas/CallSummary"
        turns:
          type: array
          items:
            $ref: "#/components/schemas/Turn"
        generations:
          type: array
          items:
            $ref: "#/components/schemas/AgentGeneration"

    Turn:
      type: object
      properties:
        turn_id:
          type: string
          format: uuid
        seq:
          type: integer
        started_at:
          type: integer
        finalized_at:
          type: integer
          nullable: true
        text_final:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        state:
          type: string

    AgentGeneration:
      type: object
      properties:
        agent_generation_id:
          type: string
          format: uuid
        turn_id:
          type: string
          format: uuid
        state:
          type: string
        route_a_label:
          type: string
          nullable: true
        route_a_confidence:
          type: number
          nullable: true
        specialist:
          type: string
          nullable: true
        created_at:
          type: integer

    # ---- WebRTC Signaling schemas ----

    CreateCallResponse:
      type: object
      required: [call_id, status]
      properties:
        call_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created]

    SDPRequest:
      type: object
      required: [sdp, type]
      properties:
        sdp:
          type: string
          description: SDP offer string
        type:
          type: string
          enum: [offer]

    SDPResponse:
      type: object
      required: [sdp, type]
      properties:
        sdp:
          type: string
          description: SDP answer string
        type:
          type: string
          enum: [answer]